Lint

+++ src/lang/AtomType.c
@@ -7,14 +7,11 @@ char const * atomTypeNames[N_ATOMTYPES + 1] = {
 	"NONE",
 	"NAME",
 	"ID",
-
 	"UINT",
 	"INT",
-	"FLOAT64",
-
+	"FLOAT",
 	"LETTER",
 	"VARIABLE",
-	
 	"PARAMETER",
 };
